requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
langchain>=0.1.0
langchain-together>=0.1.0
langchain-community>=0.0.20
langchain-experimental>=0.0.50
openai>=1.0.0
python-dotenv>=1.0.0